refactor: waiting 패키지 변경 (repository -> domain)
