+++ src/main/java/roomescape/reservation/application/dto/MyReservation.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import roomescape.reservation.domain.Reservation;
-import roomescape.waiting.Waiting;
-import roomescape.waiting.WaitingWithRank;
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.domain.WaitingWithRank;
 
 public record MyReservation(
         Long id,
