+++ src/main/java/roomescape/waiting/application/WaitingService.java
@@ -11,12 +11,12 @@ import roomescape.reservation.application.dto.MemberReservationRequest;
 import roomescape.reservation.application.dto.MyReservation;
 import roomescape.waiting.application.dto.WaitingIdResponse;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.waiting.Waiting;
-import roomescape.waiting.WaitingWithRank;
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.domain.WaitingWithRank;
 import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.reservation.domain.repository.ReservationTimeRepository;
 import roomescape.waiting.application.dto.WaitingInfoResponse;
-import roomescape.waiting.repository.WaitingRepository;
+import roomescape.waiting.domain.repository.WaitingRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.repository.ThemeRepository;
