+++ src/test/java/roomescape/waiting/WaitingRepositoryTest.java
@@ -16,7 +16,9 @@ import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.repository.ThemeRepository;
-import roomescape.waiting.repository.WaitingRepository;
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.domain.repository.WaitingRepository;
+import roomescape.waiting.domain.WaitingWithRank;
 
 @ActiveProfiles("test")
 @DataJpaTest
